LPS-67766 Move task class to "tasks" package
